refactor(naming): fixing typo in contracts folder
