+++ tech-challenge-ecosystem/src/core/contracts/messaging/fixture/transaction.message.producer.fixture.ts
