+++ tech-challenge-ecosystem/src/data/repository/repository.providers.ts
@@ -1,6 +1,6 @@
-import { TRANSACTION_DATASOURCE_PROVIDE } from 'src/core/constracts/data/transaction.datasource';
+import { TRANSACTION_DATASOURCE_PROVIDE } from 'src/core/contracts/data/transaction.datasource';
 import { TransactionRepository } from './transaction.repository';
-import { PAYABLE_DATASOURCE_PROVIDE } from 'src/core/constracts/data/payable.datasource';
+import { PAYABLE_DATASOURCE_PROVIDE } from 'src/core/contracts/data/payable.datasource';
 import { PayableRepository } from './payable.repository';
 
 export const repositoryProviders = [
